Make migration only start once the "Next" button is pressed.
